requirements-core.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements-core.in
#
aiofiles==24.1.0
    # via -r requirements-core.in
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   openai
    #   starlette
async-timeout==5.0.1
    # via redis
black==25.1.0
    # via -r requirements-core.in
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   black
    #   uvicorn
distro==1.9.0
    # via openai
exceptiongroup==1.3.0
    # via
    #   anyio
    #   pytest
fastapi==0.116.1
    # via -r requirements-core.in
google-api-core==2.25.1
    # via google-api-python-client
google-api-python-client==2.178.0
    # via -r requirements-core.in
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.2.0
    # via google-api-python-client
googleapis-common-protos==1.70.0
    # via google-api-core
gunicorn==23.0.0
    # via -r requirements-core.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   -r requirements-core.in
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
jiter==0.10.0
    # via openai
mypy-extensions==1.1.0
    # via black
numpy==2.2.6
    # via
    #   -r requirements-core.in
    #   pandas
openai==1.99.5
    # via -r requirements-core.in
packaging==25.0
    # via
    #   black
    #   gunicorn
    #   pytest
pandas==2.3.1
    # via -r requirements-core.in
pathspec==0.12.1
    # via black
platformdirs==4.3.8
    # via black
pluggy==1.6.0
    # via pytest
proto-plus==1.26.1
    # via google-api-core
protobuf==6.31.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   -r requirements-core.in
    #   fastapi
    #   openai
    #   sqlmodel
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via pytest
pyparsing==3.2.3
    # via httplib2
pytest==8.4.1
    # via -r requirements-core.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via -r requirements-core.in
python-multipart==0.0.20
    # via -r requirements-core.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via -r requirements-core.in
redis==6.4.0
    # via -r requirements-core.in
requests==2.32.4
    # via
    #   -r requirements-core.in
    #   google-api-core
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.42
    # via
    #   -r requirements-core.in
    #   sqlmodel
sqlmodel==0.0.24
    # via -r requirements-core.in
starlette==0.47.2
    # via fastapi
tomli==2.2.1
    # via
    #   black
    #   pytest
tqdm==4.67.1
    # via openai
typing-extensions==4.14.1
    # via
    #   anyio
    #   black
    #   exceptiongroup
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via -r requirements-core.in
